Move both radio button icons to a new file
