+++ LoveMyCapibara/LoveMyCapibara/Views/Scenes/Settings/Subview/SelectedRadioIcon.swift
@@ -1,0 +1,22 @@
+//
+//  SelectedRadioIcon.swift
+//  LoveMyCapibara
+//
+//  Created by userext on 01/06/23.
+//
+
+import SwiftUI
+
+struct SelectedRadioIcon: View {
+    var body: some View {
+        Image("Icon Select _ Selected")
+            .resizable()
+            .frame(width: 20, height: 20)
+    }
+}
+
+struct SelectedRadioIcon_Previews: PreviewProvider {
+    static var previews: some View {
+        SelectedRadioIcon()
+    }
+}
